requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.3
faker>=22.0.0

# Machine Learning
scikit-learn>=1.4.0
prophet>=1.1.5
joblib>=1.3.2
# mlflow (removed as we use local tracking and it's optional, but keeping minimal for compatibility if needed. Actually we used mlflow in code so we need it)
mlflow>=2.9.2

# GenAI & RAG
langchain>=1.2.10
langchain-community>=0.4.1
langchain-core>=1.2.11
langchain-openai>=1.1.9
langchain-text-splitters>=1.1.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
openai>=1.7.2
tiktoken>=0.5.2
langchain-google-genai

# Frontend
streamlit>=1.30.0
plotly>=5.18.0
streamlit-chat>=0.1.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.26.0
tenacity